Remove the ephemeral and optional native flags.
Natives are now optional by default. Native calls are always preceded by an
error check, and unbound native calls will throw an error. It is the
responsibility of the embedder to decide whether to reject scripts outright
due to having unbound natives in its native table.

However, a new IMMUTABLE native binding flag has been introduced as well. This
flag prevents any further mutation to the bindings for a native, allowing the
VM to remove any additional checks. Again, it is the responsibility of the
embedding application to decide whether or not a native can be immutable.