Replace token_pos_t with a 32-bit identifier.
This trick is borrowed from Clang. token_pos_t is both heavyweight and
inaccurate. We can be more accurate with a much smaller memory
footprint. The new method of tracking positions is via SourceLocation, a
32-bit integer. The way it works is simple.

Each time we enter a file or macro, we give it an ID. To make a
SourceLocation originating from that file or macro, we take the ID and
add the current byte offset into its character stream.

To make sure SourceLocations from different streams are unique, we keep
a global counter indicating the maximum possible SourceLocation. This
allows us to ensure that each file/macro we enter gets a new ID that
will not collide with any previous entries.

Each ID is registered in a table (which is inherently sorted) for fast
reverse lookup. This table also tracks where each stream originated
from. This allows finding the full location history of a macro.

This also reduces memory use in the AST by about 5%.